fix: 모듈 분리 수정
